+++ CHANGELOG.md
@@ -253,7 +253,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First test release.
 
-[Unreleased]: https://gitlab.com/smspp/tools/-/compare/0.6.0...develop
+[Unreleased]: https://gitlab.com/smspp/tools/-/compare/0.7.1...develop
+[0.7.1]: https://gitlab.com/smspp/tools/-/compare/0.7.0...0.7.1
+[0.7.0]: https://gitlab.com/smspp/tools/-/compare/0.6.0...0.7.0
 [0.6.0]: https://gitlab.com/smspp/tools/-/compare/0.5.4...0.6.0
 [0.5.4]: https://gitlab.com/smspp/tools/-/compare/0.5.3...0.5.4
 [0.5.3]: https://gitlab.com/smspp/tools/-/compare/0.5.2...0.5.3